[REF] spreadsheet_dashboard: separate filter value list
Before this commit:
- Dashboard search bar menu handled both filter value list state management and
  rendering.
- This coupling constrained the filter value list to the search bar menu.

After this commit:
- Filter value list is now responsible only for displaying values and
  emitting updates through props.
- State management is moved to the dashboard search bar menu.
- Confirm and discard actions are now handled in the search bar menu.

Task: 5114625
