+++ addons/spreadsheet_dashboard/static/src/bundle/dashboard_action/dashboard_filter_list/dashboard_filter_list.scss
