+++ addons/spreadsheet_dashboard/static/src/bundle/dashboard_action/dashboard_filter_list/dashboard_filter_list.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+
+    <t t-name="spreadsheet.DashboardFilterList">
+        <div class="d-flex flex-column o-filter-values w-100">
+            <div t-foreach="this.props.filtersAndValues"
+                    t-as="item"
+                    t-key="item.globalFilter.id"
+                    class="d-flex o-filter-item align-items-center"
+                    t-att-class="this.env.isSmall ? 'dropdown-item' : ''"
+                    t-att-data-id="item.globalFilter.id">
+                <div class="me-1 w-25">
+                    <t t-out="this.getTranslatedFilterLabel(item.globalFilter)"/>
+                </div>
+                <t t-set="operators" t-value="this.getOperators(item.globalFilter)"/>
+                <select t-if="operators?.length" class="o_input w-25 pe-3 me-3 text-truncate" t-on-change="(ev) => this.updateOperator(item, ev.target.value)">
+                    <t t-foreach="operators" t-as="operator" t-key="operator">
+                        <option class="text-black" t-att-value="operator" t-att-selected="item.value?.operator === operator" t-out="this.getOperatorLabel(operator)" />
+                    </t>
+                </select>
+                <div t-else="" class="w-25 me-3"/>
+                <FilterValue filter="item.globalFilter"
+                                model="this.props.model"
+                                globalFilterValue="item.value"
+                                setGlobalFilterValue="(_, value) => this.updateValue(item, value)" />
+                <div class="o-filter-clear">
+                    <button t-if="this.filterHasClearButton(item)"
+                        class="btn btn-link px-2 text-danger fs-4"
+                        t-on-click="() => this.clearFilter(item)">
+                        <i class="fa fa-times"></i>
+                    </button>
+                </div>
+            </div>
+        </div>
+    </t>
+
+</templates>
