+++ addons/spreadsheet_dashboard/static/src/bundle/dashboard_action/dashboard_search_bar_menu/dashboard_search_bar_menu.js
@@ -1,24 +1,21 @@
 import { useState } from "@web/owl2/utils";
 import { Component, onWillStart } from "@odoo/owl";
-import { FilterValue } from "@spreadsheet/global_filters/components/filter_value/filter_value";
-import { _t } from "@web/core/l10n/translation";
-import { getOperatorLabel } from "@web/core/tree_editor/tree_editor_operator_editor";
 import {
     getDefaultValue,
     getEmptyFilterValue,
-    getFilterTypeOperators,
     isEmptyFilterValue,
 } from "@spreadsheet/global_filters/helpers";
 import { useService } from "@web/core/utils/hooks";
 import { deepEqual } from "@web/core/utils/objects";
+import { DashboardFilterList } from "../dashboard_filter_list/dashboard_filter_list";
 
 /**
- * This component is used to display a list of all the global filters of a dashboard.
- * It allows the user to select the values of the filters and confirm or discard the changes.
+ * This component manages the state and behavior of the filter value list
+ * in the dashboard search bar.
  */
 export class DashboardSearchBarMenu extends Component {
     static template = "spreadsheet_dashboard.DashboardSearchBarMenu";
-    static components = { FilterValue };
+    static components = { DashboardFilterList };
 
     static props = {
         close: Function,
@@ -45,64 +42,20 @@ export class DashboardSearchBarMenu extends Component {
         return this.props.model.getters.getGlobalFilters();
     }
 
-    setGlobalFilterValue(node, value) {
-        if (value == undefined && node.globalFilter.type !== "date") {
-            // preserve the operator.
-            node.value = {
-                ...node.value,
-                ...getEmptyFilterValue(node.globalFilter, node.value.operator),
-            };
-        } else {
-            node.value = value;
-        }
-    }
-
-    getTranslatedFilterLabel(filter) {
-        return _t(filter.label); // Label is extracted from the spreadsheet json file
-    }
-
-    getOperators(filter) {
-        const operators = getFilterTypeOperators(filter.type);
-        if (filter.type === "relation" && !this.searchableParentRelations[filter.modelName]) {
-            return operators.filter((op) => op !== "child_of");
-        }
-        return filter.type === "boolean" ? [undefined, ...operators] : operators;
-    }
-
-    filterHasClearButton(node) {
-        return !isEmptyFilterValue(node.globalFilter, node.value);
-    }
-
-    getOperatorLabel(operator) {
-        return operator ? getOperatorLabel(operator) : "";
-    }
-
-    updateOperator(node, operator) {
-        if (!operator) {
-            node.value = undefined;
+    onFilterChange(filterId, value) {
+        const node = this.state.filtersAndValues.find((f) => f.globalFilter.id === filterId);
+        if (!node) {
             return;
         }
-        if (!node.value) {
-            node.value = {};
-        }
-        node.value.operator = operator;
-        const defaultValue = getEmptyFilterValue(node.globalFilter, operator);
-        for (const key of Object.keys(defaultValue ?? {})) {
-            if (!(key in node.value)) {
-                node.value[key] = defaultValue[key];
-            }
-        }
-    }
-
-    clearFilter(filterId) {
-        const node = this.state.filtersAndValues.find((node) => node.globalFilter.id === filterId);
-        if (node && node.value) {
+        if (value === undefined && node.value?.operator) {
             const emptyValue = getEmptyFilterValue(node.globalFilter, node.value.operator);
             node.value =
                 typeof emptyValue === "object"
                     ? { ...emptyValue, operator: node.value.operator }
                     : emptyValue;
+            return;
         }
+        node.value = value;
     }
 
     onConfirm() {
