+++ addons/spreadsheet_dashboard/static/src/bundle/dashboard_action/dashboard_search_bar_menu/dashboard_search_bar_menu.xml
@@ -2,35 +2,12 @@
 <templates xml:space="preserve">
 
     <t t-name="spreadsheet_dashboard.DashboardSearchBarMenu">
-        <div class="d-flex flex-column o-filter-values">
-            <div t-foreach="this.state.filtersAndValues"
-                    t-as="item"
-                    t-key="item_index"
-                    class="d-flex o-filter-item align-items-center"
-                    t-att-class="this.env.isSmall ? 'dropdown-item' : ''"
-                    t-att-data-id="item.globalFilter.id">
-                <div class="me-1 w-25">
-                    <t t-out="this.getTranslatedFilterLabel(item.globalFilter)"/>
-                </div>
-                <t t-set="operators" t-value="this.getOperators(item.globalFilter)"/>
-                <div t-if="!operators?.length" class="w-25 me-3"/>
-                <select t-else="" class="o_input w-25 pe-3 me-3 text-truncate" t-on-change="(ev) => this.updateOperator(item, ev.target.value)">
-                    <t t-foreach="operators" t-as="operator" t-key="operator">
-                        <option class="text-black" t-att-value="operator" t-att-selected="item.value?.operator === operator" t-out="this.getOperatorLabel(operator)" />
-                    </t>
-                </select>
-                <FilterValue filter="item.globalFilter"
-                                model="this.props.model"
-                                setGlobalFilterValue="(id, value) => this.setGlobalFilterValue(item, value)"
-                                globalFilterValue="item.value"/>
-                <div class="o-filter-clear">
-                    <button t-if="this.filterHasClearButton(item)"
-                        class="btn btn-link px-2 text-danger fs-4"
-                        t-on-click="() => this.clearFilter(item.globalFilter.id)">
-                        <i class="fa fa-times"></i>
-                    </button>
-                </div>
-            </div>
+        <div class="d-flex">
+            <DashboardFilterList
+                filtersAndValues="this.state.filtersAndValues"
+                searchableParentRelations="this.searchableParentRelations"
+                onFilterChange.bind="this.onFilterChange"
+                model="this.props.model"/>
         </div>
         <div class="o-filter-values-footer d-flex mt-3 pt-3 border-top gap-1">
             <button class="btn btn-primary" t-on-click="this.onConfirm">Filter</button>
