+++ addons/spreadsheet_dashboard/static/tests/dashboard/dashboard_action.test.js
@@ -7,7 +7,12 @@ import {
     defineSpreadsheetDashboardModels,
     getDashboardServerData,
 } from "@spreadsheet_dashboard/../tests/helpers/data";
-import { contains, onRpc, patchWithCleanup } from "@web/../tests/web_test_helpers";
+import {
+    contains,
+    onRpc,
+    patchWithCleanup,
+    toggleSearchBarMenu,
+} from "@web/../tests/web_test_helpers";
 import { browser } from "@web/core/browser/browser";
 import { RPCError } from "@web/core/network/rpc";
 import { Deferred } from "@web/core/utils/concurrency";
@@ -730,5 +735,223 @@ describe("Quick search bar", () => {
 
         filterValue = model.getters.getGlobalFilterValue(productFilter.id);
         expect(filterValue).toEqual(undefined);
+    });
+});
+
+describe("Filter list behavior in search bar", () => {
+    async function setupDashboardWithFilter(globalFilter) {
+        const spreadsheetData = { globalFilters: [globalFilter] };
+        const serverData = getServerData(spreadsheetData);
+        const { model } = await createSpreadsheetDashboard({ serverData });
+        await toggleSearchBarMenu();
+        return { model };
+    }
+
+    test("Can set a text filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "text",
+            label: "Text Filter",
+        });
+
+        await contains(".o-filter-values select").select("not ilike");
+        await contains(".o-filter-values .o-autocomplete input").edit("foo");
+        await press("Enter");
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined);
+
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual({
+            operator: "not ilike",
+            strings: ["foo"],
+        });
+    });
+
+    test("Can set a numeric filter value with basic operator", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "numeric",
+            label: "Numeric Filter",
+        });
+
+        await contains(".o-filter-values select").select(">");
+        await contains(".o-filter-values input").edit(1998);
+        await press("Enter");
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined, {
+            message: "value is not directly set",
+        });
+
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual(
+            { operator: ">", targetValue: 1998 },
+            { message: "value is set" }
+        );
+    });
+
+    test("Can set a numeric filter value with between operator", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "numeric",
+            label: "Numeric Filter",
+        });
+
+        await contains(".o-filter-values select").select("between");
+        const [minInput, maxInput] = document.querySelectorAll(".o-global-filter-numeric-value");
+        expect([minInput, maxInput]).toHaveLength(2);
+        await contains(minInput).edit(1);
+        await contains(maxInput).edit(99);
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual(
+            { operator: "between", minimumValue: 1, maximumValue: 99 },
+            { message: "value is set" }
+        );
+    });
+
+    test("Can set a relation filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "relation",
+            modelName: "product",
+            label: "Relation Filter",
+        });
+
+        await contains(".o-filter-values select").select("not in");
+        await contains("input.o-autocomplete--input").click();
+        await contains(".o-autocomplete--dropdown-item:first").click();
+        expect(".o_tag").toHaveCount(1);
+
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual(
+            { operator: "not in", ids: [37] },
+            { message: "value is set" }
+        );
+
+        await toggleSearchBarMenu();
+        await contains(".o_tag .o_delete").click();
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined);
+    });
+
+    test("Can remove a default relation filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "relation",
+            modelName: "product",
+            label: "Relation Filter",
+            defaultValue: { operator: "in", ids: [37] },
+        });
+        expect(".o_tag").toHaveCount(1);
+
+        await contains(".o_tag .o_delete").click();
+        expect(".o_tag").toHaveCount(0);
+
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined);
+    });
+
+    test("Can change a boolean filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "boolean",
+            label: "Boolean Filter",
+        });
+        expect(".o-filter-values select").toHaveValue("");
+
+        await contains(".o-filter-values select").select("not set");
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual({ operator: "not set" });
+
+        await toggleSearchBarMenu();
+        await contains(".o-filter-values select").select("");
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined);
+    });
+
+    test("Can set a date filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "date",
+            label: "Date Filter",
+        });
+
+        await contains(".o-date-filter-input").click();
+        await contains(".o-dropdown-item[data-id='last_7_days']").click();
+        expect(".o-date-filter-input").toHaveValue("Last 7 Days");
+
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual({
+            period: "last_7_days",
+            type: "relative",
+        });
+    });
+
+    test("Readonly user can update a filter value", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "text",
+            label: "Text Filter",
+        });
+        model.updateMode("readonly");
+
+        await contains(".o-filter-values .o-autocomplete input").edit("foo");
+        await press("Enter");
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42").strings).toEqual(["foo"], {
+            message: "value is set",
+        });
+    });
+
+    test("Can clear a filter value removing the values manually", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "text",
+            label: "Text Filter",
+            defaultValue: { operator: "ilike", strings: ["foo"] },
+        });
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(1);
+
+        await contains(".o-filter-values .o-filter-item .o_tag .o_delete").click();
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(0);
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined, {
+            message: "value is cleared",
+        });
+    });
+
+    test("Can clear a filter value with the clear button", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "text",
+            label: "Text Filter",
+            defaultValue: { operator: "ilike", strings: ["foo"] },
+        });
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(1);
+
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(1);
+        await contains(".o-filter-values .o-filter-item .o-filter-clear button").click();
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(0);
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toBe(undefined);
+    });
+
+    test("clearing a filter value preserves the operator", async function () {
+        const { model } = await setupDashboardWithFilter({
+            id: "42",
+            type: "text",
+            label: "Text Filter",
+            defaultValue: { operator: "ilike", strings: ["foo"] },
+        });
+
+        await contains(".o-filter-values select").select("starts with");
+        await contains(".o-filter-values .o-filter-item .o_tag .o_delete").click();
+        expect(".o-filter-values .o-filter-item .o_tag").toHaveCount(0);
+        expect(".o-filter-values select").toHaveValue("starts with");
+
+        await contains(".o-filter-values .o-filter-item .o-autocomplete input").edit("foo");
+        await contains(".o-filter-values .o-filter-item .o-autocomplete input").press("Enter");
+        await contains(".o-filter-values-footer .btn-primary").click();
+        expect(model.getters.getGlobalFilterValue("42")).toEqual({
+            operator: "starts with",
+            strings: ["foo"],
+        });
     });
 });
